+++ djangocms_frontend/contrib/tabs/frameworks/bootstrap5.py
@@ -10,12 +10,12 @@ class TabItemRenderMixin:
         parent = instance.parent.get_plugin_instance()[0]
         instance.add_classes("tab-pane", parent.tab_effect)
         if instance.config.get("tab_bordered", False):
-            if context["parent"].tab_type == "nav-tabs":
+            if parent.tab_type == "nav-tabs":
                 instance.add_classes("border-end border-start border-bottom")
             else:
                 instance.add_classes("border")
-        if context["parent"].tab_type == "nav-tabs":
+        if parent.tab_type == "nav-tabs":
             instance.add_classes("bg-body")
-        if parent.tab_index == context["parentloop"]["counter"]:
+        if parent.tab_index == context.get("parentloop", {}).get("counter"):
             instance.add_classes("show active")
         return super().render(context, instance, placeholder)